+++ Assets/Scripts/ServerContainer.cs
@@ -1,5 +1,4 @@
 using JamForge.Resolver;
-using MessagePipe;
 using VContainer;
 
 public class TestServices
@@ -11,7 +10,6 @@ public class ServerContainer : JamInstaller
 {
     public override void Install(IContainerBuilder builder)
     {
-        builder.RegisterMessageBroker<GameStartMessage>(MessagePipeOptions);
         builder.Register<TestServices>(Lifetime.Singleton);
     }
 }