+++ Packages/com.anomyx.jamforge/Runtime/Facade/JamFacade_Messages.cs
@@ -1,0 +1,17 @@
+using JamForge.Messages;
+using VContainer;
+
+namespace JamForge
+{
+    public partial class Jam
+    {
+        #region Events
+
+        [Inject]
+        private IMessageCenter _messageCenter;
+
+        public static IMessageCenter Messages => Instance._messageCenter;
+
+        #endregion
+    }
+}