+++ Packages/com.anomyx.jamforge/Runtime/Services/Messages/Abstractions/IMessage.cs
@@ -1,0 +1,4 @@
+namespace JamForge.Messages
+{
+    public interface IMessage { }
+}