+++ Packages/com.anomyx.jamforge/Runtime/Services/Resolver/Implements/JamInstaller.cs
@@ -1,4 +1,3 @@
-using MessagePipe;
 using VContainer;
 using VContainer.Unity;
 
@@ -6,8 +5,6 @@ namespace JamForge.Resolver
 {
     public abstract class JamInstaller : LifetimeScope
     {
-        protected MessagePipeOptions MessagePipeOptions { get; set; } = new();
-
         public abstract void Install(IContainerBuilder builder);
 
         protected sealed override void Configure(IContainerBuilder builder)